Loading .env even if APP_ENV is set
This is tough: we're getting WTF moments when someone tries to set,
for example, APP_ENV=test or some other environment, and suddenly
the .env file isn't loaded.

This is to hopefully make things work in a more predictable way: .env
is always loaded, except in the prod environment. This works because
DotEnv does not override any existing variables, so there's no risk
in loading it.