Load routes for specific env first
It's common to add a 'wildcard' route to an application that looks like
a CMS. Usually the route has the following path `/{slug}`.

But with the current routes loading order, the 'dev' routing is loaded
after regular routes. In previous SF versions, dev routing was loaded
first. This PR retores this behavior.