Check for the dev environment instead of prod
It feels dangerous that we're checking whether the environment !== 'prod'.

I think APP_DEBUG should only default to true if the environment is specifically "dev".

In a real example I have created a "staging" environment and it is showing debug information. I now need to specifically create a env var to stop this. Unless users are aware of this (or there is human error in creating this env var), the server could reveal too much information creating security issues.