Replaced silent fail for loading php env file
Replaced with actual test if file exists since the suppressing of an error just causes issues.

For instance if you run Xdebug.halt_level = E_WARNING

Xdebug notices that an error has happened but since we silenty fail for the include we never get an error output. With the change in place we do see an error if we want to see it.