Align code coverage settings with documentation
After noticing slow coverage collection on our jenkins server for our symfony project, i noticed after some time that we were missing the on the whitelist configuration the processUncoveredFilesFromWhitelist attribute. I tried to backtrace why this configuration was missing and realized that this might come from the phpunit flex recipe. So here we go with the fix.

Fixes issues with slow coverage collection mentioned here:  
https://github.com/sebastianbergmann/phpunit/issues/2489  
Also the documentation already includes the configuration:  
https://symfony.com/doc/4.4/create_framework/unit_testing.html
https://symfony.com/doc/current/create_framework/unit_testing.html