Put doctrine proxies in build dir for production
When auto_generate_proxy_classes: false is set, the proxies need to be pre-generated with the cache warmer. So in this case, the files are much better suited in the build dir. The build dir is the same as the cache dir by default. But for people trying to use the build dir properly, this makes much more sense. See https://github.com/symfony/symfony/issues/40827#issuecomment-820653578